About and Media page
About page - I added the pdf section of the about page to the CMS, and
added the photo date.

Media page - I removed the download attribute to the gallery links,
because they weren’t helping at all. I’ll have to re-visit this at a
later date.

CSS - reduced font-size of they lyrics section, and added a media query
to the pdf section of the about page.