Functions and WooCommerce
I removed js files from functions.php that were no longer needed, and
removed a piece of code that added the production variation name to the
message when adding a product to the cart, because the functions were
deprecated after the new WooCommerce update to 3.0.

I also added cart-totals.php to the theme template files, because Mark
requested that “Totals” be capitalized.